issue #25: Rework timestamp corruption patch to key off kansa_charge rather than earliest_charge
It was messing with payment timestamps that weren't imported, and this
was creating artifacts in the export. Narrowing down to just the kansa
payment fixes this.

You could only have one payment for your Kansa purchase, so this is
safe to assume #find will work. I only saw 2 artifacts anyway.