maint: Squash auto load constant warning
This would have become un-bootable in future versions of rails. Warning
was as follows:

DEPRECATION WARNING: Initialization autoloaded the constant SetHugoGlobals.

Being able to do this is deprecated. Autoloading during initialization is going
to be an error condition in future versions of Rails.

Reloading does not reboot the application, and therefore code executed during
initialization does not run again. So, if you reload SetHugoGlobals, for example,
the expected changes won't be reflected in that stale Class object.

`config.autoloader` is set to `classic`. This autoloaded constant would have been unloaded if `config.autoloader` had been set to `:zeitwerk`.

Please, check the "Autoloading and Reloading Constants" guide for solutions.
 (called from <top (required)> at /Users/mbgray/code/conzealand/config/environment.rb:21)