maint: Factory reservations create charges by default
Before it was executing the before(:create) hooks in an order that
stopped it creating charges outside of a very narrow set of conditions.

Now we create memberships and users, then go on to try making charges.

This change also makes explicit when we want a membership in instalment
without any charges.