maint: Stop running linting rules against master and tags
This was done because the github rules for rubocop are not maintained
enough for us to use ruby 2.7 which we needed to move to in order to
stay current and security patched.

As we've changed our linting rules, this means a lot of files are
running invalid versions of our current rules.

So this stage in our pipeline is really here to enforce files that are
being changed to pass the current linting rules before they can be
merged.