Disable rubocop on pre-commit
While I see the value or rubocop, it's literally a minute to every commit. I'd
buy the value of it if we found a way to hook the precommit hook up to a project
rubocop-daemon, but holy gods does the current implementation ever
hurt (literally! My laptop heats up enough running it that I need to hold it off
my lap! 😂)