Extract theme/contact selection into a concern
Rather than reference the environment whenever we need a theme or contact model,
this loads those values from the environment and then pushes them down into the
system via a ThemeConcern.

This differs from optional values like hugo ranges in that a con *must* have a
theme and contact model, so this forces the selection early in the process by
bailing in the rails application config stage if one is not set.