maint: Sync to gloo from Claim, Charge, Order or ConzealandContact
These are all the places where state changes happen, such as
- Update contact details for a membership
- Create a new claim on a membership
- Upgrade a membership
- Putting in an instalment
- Putting in a final payment to pay off a membership
- Transfer a membership between two users

It still feels right to have this in the models because it's going to be
something that needs to be tracked by an external system.

Yes this is buckshot because that external system doesn't model the same
as ours, so this is the best we can do.