feat: add support for non-public membership types
Some memberships should be possible to purchase, but not displayed in
the set of available memberships. If a direct link to a membership is
provided, it will still be purchaseable, but it won't be in the default
list.

If the support user is logged in, it would be nice to have these show
up, but with a "private" indicator and a "get link" option instead of a
"buy now" option; something that copies the link to the clipboard.

Fixes #101