Update nominations_reminder_3_days_left.html.erb
Hi Chris - this will only go out to people who have nomination rights, right?

I'm not sure how to tell what the subject line will be for this email. I don't see it in this code. Can we make sure it doesn't mention Retro Hugos and that the year is right?

And just want to confirm the variables are set up for this year's Hugos:
@hugo_nom_deadline = 11:59 pm PDT, 15 March 2022 
@member_login_url = https://registration.chicon.org/
@hugo_ballot_pub_month = April
@worldcon_public_name = Chicon 8
@start_day_informal = September 1
@end_day_informal = September 5
@worldcon_year = 2022
@worldcon_year_before = 2021
@email_hugo_help = hugo-help@chicon.org
@organizers_names_for_signature = Kat Jones and Nicholas Whyte